yolo-service/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.20.0
numpy==1.24.3
opencv-python==4.7.0.72
python-multipart==0.0.6
Pillow==9.5.0